+++ app/src/main/java/com/techcourse/DispatcherServletInitializer.java
@@ -5,6 +5,7 @@ import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import web.org.springframework.web.WebApplicationInitializer;
+import webmvc.org.springframework.web.servlet.mvc.DispatcherServlet;
 import webmvc.org.springframework.web.servlet.mvc.HandlerAdapters;
 import webmvc.org.springframework.web.servlet.mvc.HandlerMappings;
 import webmvc.org.springframework.web.servlet.mvc.tobe.AnnotationHandlerAdapter;
